Moved and renamed most of the basic processing functions
