--- pycomlink/processing/A_R_relation/__init__.py
@@ -1,10 +1,0 @@
-# ----------------------------------------------------------------------------
-# Name:
-# Purpose:
-#
-# Authors:
-#
-# Created:
-# Copyright:    (c) Christian Chwala 2014
-# Licence:      The MIT License
-# ----------------------------------------------------------------------------
